Fix function signatures

+++ inc/file.h
@@ -80,7 +80,7 @@ namespace rtm {
 	int64_t	fileReaderSeek(FileReaderHandle _handle, int64_t _offset, uint64_t _origin = File::Seek_CUR);
 
 	///
-	int32_t	fileReaderRead(FileReaderHandle _handle, void* _dest, int64_t _size);
+	int64_t	fileReaderRead(FileReaderHandle _handle, void* _dest, int64_t _size);
 
 	///
 	int64_t	fileReaderGetSize(FileReaderHandle _handle);
@@ -108,7 +108,7 @@ namespace rtm {
 	int64_t	fileWriterSeek(FileWriterHandle _handle, int64_t _offset, uint64_t _origin = File::Seek_CUR);
 
 	///
-	int32_t	fileWriterWrite(FileWriterHandle _handle, void* _src, int64_t _size);
+	int64_t	fileWriterWrite(FileWriterHandle _handle, void* _src, int64_t _size);
 
 	/// <summary>
 	int64_t	fileWriterGetSize(FileWriterHandle _handle);
